# Conflicts:
#	docs/source/mpsoc/ip_cores/ip_cores.rst
#	ip_cores/vv_index.xml
#	javascope/properties.ini
#	vitis/software/Baremetal/src/uz/default_uz_global_configuration.h
#	vitis/vivado_exported_xsa/zusys_wrapper.xsa
#	vivado/DebugProbes.ltx
#	vivado/constraints/Digital_AdapterBoard_D4.xdc
#	vivado/constraints/Digital_AdapterBoard_D5.xdc
#	vivado/project/ultrazohm.xpr
#	vivado/project/zusys/hdl/zusys_wrapper.vhd
#	vivado/zusys/zusys.bd

+++ vitis/software/Baremetal/src/IP_Cores/uz_myIP2/uz_myIP2.c
@@ -11,7 +11,7 @@ struct uz_myIP2_t
     struct uz_myIP2_config_t config;
 };
 
-static size_t instance_counter = 0U;
+static uint32_t instance_counter = 0U;
 static uz_myIP2_t instances[UZ_MYIP2_MAX_INSTANCES] = {0};
 
 static uz_myIP2_t *uz_myIP2_allocation(void);
